+++ src/Operation/Terminating/Reduce.php
@@ -26,6 +26,6 @@ final class Reduce extends FinalOperation
     
     public function handle(Signal $signal): void
     {
-        $this->reducer->consume($signal->item->value);
+        $this->reducer->consume($signal->item->value, $signal->item->key);
     }
 }