+++ src/Reducer/Average.php
@@ -13,9 +13,10 @@ final class Average implements Reducer
     
     /**
      * @param float|int $value
+     * @param string|int|null $key
      * @return void
      */
-    public function consume($value): void
+    public function consume($value, $key = null): void
     {
         $this->total += $value;