+++ src/Reducer/GenericReducer.php
@@ -16,16 +16,15 @@ final class GenericReducer implements Reducer
     private bool $isFirst = true;
     private bool $hasAny = false;
     
+    private int $numOfArgs;
+    
     public function __construct(callable $reducer)
     {
-        if (Helper::getNumOfArgs($reducer) !== 2) {
-            throw new \UnexpectedValueException('Reducer have to accept 2 arguments');
-        }
-    
         $this->reducer = $reducer;
+        $this->numOfArgs = Helper::getNumOfArgs($reducer);
     }
     
-    public function consume($value): void
+    public function consume($value, $key = null): void
     {
         $this->hasAny = true;
         
@@ -34,7 +33,12 @@ final class GenericReducer implements Reducer
             $this->result = $value;
         } else {
             $reduce = $this->reducer;
-            $this->result = $reduce($this->result, $value);
+            switch ($this->numOfArgs) {
+                case 2: $this->result = $reduce($this->result, $value); break;
+                case 3: $this->result = $reduce($this->result, $value, $key); break;
+                default:
+                    throw Helper::wrongNumOfArgsException('Reducer', $this->numOfArgs, 2, 3);
+            }
         }
     }