+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All important changes to `fiisoft-jackdaw-stream` will be documented in this file
 
+## 2.2.0 - incompatible changes!
+
+- method Reducer::consume changed, key is passed as second argument
+- added method StreamApi::moveTo
+
 ## 2.1.0 - incompatible changes!
 
 - removed method Result::__toString